refactor: move subtitle model to new file
